requirements.txt
aiogram==3.4.1
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
pydantic==2.5.3
python-dotenv==1.0.1
httpx==0.27.0